+++ common/src/runners/mod.rs
@@ -1,2 +1,2 @@
 pub mod compute_runner;
-pub mod verifiy_runner;
+pub mod verification_runner;
